# apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MPC 2-of-2 signature escrow wallet. Host (server) endpoints\nhandle auth, pairing, mailbox, keygen-session arbitration,\nescrow and timebox. Local (client) endpoints handle keygen,\naccounts, exchanges, balance, transactions and withdrawals.\nServer endpoints under /v1 (except /auth) require a Bearer JWT.",
        "title": "Signature Escrow API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/pair/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "The partner accepts a pending pair by its ID. Idempotent: returns the pair if it is already accepted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pair"
                ],
                "summary": "Accept a pair",
                "parameters": [
                    {
                        "description": "Pair ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.PairAcceptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PairAcceptResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/pair/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a pending pair request to another ETH address. Idempotent: returns the existing pair if present.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pair"
                ],
                "summary": "Create a pair",
                "parameters": [
                    {
                        "description": "Partner address",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.PairCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PairCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/pair/delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a pair entirely (both participants lose it). Only a member of the pair may delete it. Idempotent: returns success if the pair is already gone.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pair"
                ],
                "summary": "Delete a pair",
                "parameters": [
                    {
                        "description": "Pair ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.PairAcceptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/pair/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all pairs involving the caller, split into incoming (caller is partner) and outgoing (caller is initiator).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pair"
                ],
                "summary": "List pending pairs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PairPendingResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "server.Pair": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "initiator": {
                    "type": "string"
                },
                "partner": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "server.PairAcceptRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "server.PairAcceptResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "initiator": {
                    "type": "string"
                },
                "partner": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "server.PairCreateRequest": {
            "type": "object",
            "properties": {
                "partner": {
                    "type": "string"
                }
            }
        },
        "server.PairCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "initiator": {
                    "type": "string"
                },
                "partner": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "server.PairPendingResponse": {
            "type": "object",
            "properties": {
                "incoming": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.Pair"
                    }
                },
                "outgoing": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.Pair"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT obtained from /v1/auth/login, sent as \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}